Move `CreateResponse` command
