+++ wcfsetup/install/files/lib/command/comment/response/CreateResponse.class.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace wcf\system\comment\response\command;
+namespace wcf\command\comment\response;
 
 use wcf\data\comment\Comment;
 use wcf\data\comment\CommentEditor;
@@ -20,7 +20,7 @@ use wcf\system\moderation\queue\ModerationQueueActivationManager;
  * @author      Marcel Werk
  * @copyright   2001-2024 WoltLab GmbH
  * @license     GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
- * @since       6.1
+ * @since       6.2
  */
 final class CreateResponse
 {
@@ -30,8 +30,7 @@ final class CreateResponse
         private readonly ?User $user = null,
         private readonly string $username = '',
         private readonly bool $isDisabled = false,
-    ) {
-    }
+    ) {}
 
     public function __invoke(): CommentResponse
     {
